migrate db to head on start + get version
